Fixed html image paste
